requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
flask>=3.0.0

# Optional: for JS-rendered sites (e.g. bensbites.com). Install then run: playwright install
# playwright>=1.40.0
